Restructure project to allow flow-copy-source
This lets us export flow types to other projects
