+++ src/crypto/crypto.test.js
@@ -1,5 +1,5 @@
 /* global describe, it */
-import { makeFakeIos } from '../index.js'
+import { makeFakeIos } from '../indexABC.js'
 import { base16, base64, utf8 } from '../util/encoding.js'
 import { decrypt, encrypt, hmacSha256, sha256 } from './crypto.js'
 import { scrypt } from './scrypt.js'
